Move video/video_highlight tests to the new folder structure
